+++ src/actions/ens/getEnsText.ts
@@ -22,6 +22,7 @@ import {
 } from '../../utils/chain/getChainContractAddress.js'
 import { type ToHexErrorType, toHex } from '../../utils/encoding/toHex.js'
 import { isNullUniversalResolverError } from '../../utils/ens/errors.js'
+import { localBatchGatewayUrl } from '../../utils/ens/localBatchGatewayRequest.js'
 import { type NamehashErrorType, namehash } from '../../utils/ens/namehash.js'
 import {
   type PacketToBytesErrorType,
@@ -91,29 +92,27 @@ export type GetEnsTextErrorType =
  */
 export async function getEnsText<chain extends Chain | undefined>(
   client: Client<Transport, chain>,
-  {
-    blockNumber,
-    blockTag,
-    name,
-    key,
-    gatewayUrls,
-    strict,
-    universalResolverAddress: universalResolverAddress_,
-  }: GetEnsTextParameters,
+  parameters: GetEnsTextParameters,
 ): Promise<GetEnsTextReturnType> {
-  let universalResolverAddress = universalResolverAddress_
-  if (!universalResolverAddress) {
-    if (!client.chain)
+  const { blockNumber, blockTag, key, name, gatewayUrls, strict } = parameters
+  const { chain } = client
+
+  const universalResolverAddress = (() => {
+    if (parameters.universalResolverAddress)
+      return parameters.universalResolverAddress
+    if (!chain)
       throw new Error(
         'client chain not configured. universalResolverAddress is required.',
       )
-
-    universalResolverAddress = getChainContractAddress({
+    return getChainContractAddress({
       blockNumber,
-      chain: client.chain,
+      chain,
       contract: 'ensUniversalResolver',
     })
-  }
+  })()
+
+  const tlds = chain?.ensTlds
+  if (tlds && !tlds.some((tld) => name.endsWith(tld))) return null
 
   try {
     const readContractParameters = {
@@ -127,6 +126,7 @@ export async function getEnsText<chain extends Chain | undefined>(
           functionName: 'text',
           args: [namehash(name), key],
         }),
+        gatewayUrls ?? [localBatchGatewayUrl],
       ],
       blockNumber,
       blockTag,
@@ -134,12 +134,7 @@ export async function getEnsText<chain extends Chain | undefined>(
 
     const readContractAction = getAction(client, readContract, 'readContract')
 
-    const res = gatewayUrls
-      ? await readContractAction({
-          ...readContractParameters,
-          args: [...readContractParameters.args, gatewayUrls],
-        })
-      : await readContractAction(readContractParameters)
+    const res = await readContractAction(readContractParameters)
 
     if (res[0] === '0x') return null