+++ src/actions/public/getTransactionCount.test.ts
@@ -51,7 +51,7 @@ test('args: blockTag', async () => {
       address: '0xa5cc3c03994db5b0d9a5eedd10cabab0813678ac',
       blockTag: 'latest',
     }),
-  ).toBe(662)
+  ).toBe(676)
 })
 
 test('no count', async () => {