+++ src/experimental/eip5792/decorators/eip5792.test.ts
@@ -1,3 +1,5 @@
+// TODO(v3): Remove this.
+
 import { describe, expect, test } from 'vitest'
 
 import { wagmiContractConfig } from '~test/src/abis.js'
@@ -88,6 +90,8 @@ describe('smoke test', () => {
   test('getCallsStatus', async () => {
     expect(await client.getCallsStatus({ id: '0x123' })).toMatchInlineSnapshot(`
       {
+        "atomic": false,
+        "chainId": undefined,
         "receipts": [
           {
             "blockHash": "0x66a7b39a0c4635c2f30cd191d7e1fb0bd370c11dd93199f236c5bdacfc9136b3",
@@ -98,7 +102,9 @@ describe('smoke test', () => {
             "transactionHash": "0x66a7b39a0c4635c2f30cd191d7e1fb0bd370c11dd93199f236c5bdacfc9136b3",
           },
         ],
-        "status": "CONFIRMED",
+        "status": "success",
+        "statusCode": 200,
+        "version": "2.0.0",
       }
     `)
   })
@@ -110,7 +116,11 @@ describe('smoke test', () => {
         calls: [{ to: '0x0000000000000000000000000000000000000000' }],
         chain: mainnet,
       }),
-    ).toMatchInlineSnapshot(`"0x1"`)
+    ).toMatchInlineSnapshot(`
+      {
+        "id": "0x1",
+      }
+    `)
   })
 
   test('writeContracts', async () => {
@@ -124,6 +134,10 @@ describe('smoke test', () => {
           { ...wagmiContractConfig, functionName: 'mint' },
         ],
       }),
-    ).toMatchInlineSnapshot(`"0x1"`)
+    ).toMatchInlineSnapshot(`
+      {
+        "id": "0x1",
+      }
+    `)
   })
 })